fix: cast SUM(size) to bigint in quota sync to avoid PrismaClientValidationError
Postgres SUM over a BigInt column returns numeric, which Prisma maps to a
string. Passing that string into the BigInt usedSpace field threw
PrismaClientValidationError. Casting the aggregate to ::bigint yields a real
JS BigInt.

+++ backend/src/common/quota-sync.service.ts
@@ -25,7 +25,7 @@ export class QuotaSyncService {
       const results = await this.prisma.$queryRaw<
         { userId: string; actual_used: bigint }[]
       >`
-        SELECT "userId", COALESCE(SUM(size), 0) as actual_used
+        SELECT "userId", COALESCE(SUM(size), 0)::bigint as actual_used
         FROM "FileRecord"
         WHERE "deletedAt" IS NULL AND status = 'complete'
         GROUP BY "userId"
